--- src/main/java/net/bndy/wf/OAuth2TokenEnhancer.java
@@ -1,28 +1,0 @@
-package net.bndy.wf;
-
-import net.bndy.wf.modules.app.models.User;
-import org.springframework.security.oauth2.common.DefaultOAuth2AccessToken;
-import org.springframework.security.oauth2.common.OAuth2AccessToken;
-import org.springframework.security.oauth2.provider.OAuth2Authentication;
-import org.springframework.security.oauth2.provider.token.TokenEnhancer;
-
-import java.util.HashMap;
-import java.util.Map;
-
-public class OAuth2TokenEnhancer implements TokenEnhancer {
-
-    @Override
-    public OAuth2AccessToken enhance(OAuth2AccessToken oAuth2AccessToken, OAuth2Authentication oAuth2Authentication) {
-        User user = (User) oAuth2Authentication.getPrincipal();
-        final Map<String, Object> additionalInfo = new HashMap<>();
-
-        additionalInfo.put("id", user.getId());
-        additionalInfo.put("username", user.getUsername());
-        additionalInfo.put("avatar", user.getAvatar());
-        additionalInfo.put("authorities", user.getAuthorities());
-
-        ((DefaultOAuth2AccessToken) oAuth2AccessToken).setAdditionalInformation(additionalInfo);
-
-        return oAuth2AccessToken;
-    }
-}
